Static pod management moved to `Botanist` package
