+++ pkg/gardenlet/operation/botanist/staticpods.go
@@ -5,34 +5,26 @@
 package botanist
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"io"
-	"net/http"
-	"os"
 	"path/filepath"
-	"time"
 
-	"github.com/Masterminds/semver/v3"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	componentbaseconfigv1alpha1 "k8s.io/component-base/config/v1alpha1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/gardener/gardener/imagevector"
+	v1beta1helper "github.com/gardener/gardener/pkg/api/core/v1beta1/helper"
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
-	"github.com/gardener/gardener/pkg/client/kubernetes"
 	bootstrapetcd "github.com/gardener/gardener/pkg/component/etcd/bootstrap"
 	etcdconstants "github.com/gardener/gardener/pkg/component/etcd/etcd/constants"
-	resourcemanagerconstants "github.com/gardener/gardener/pkg/component/gardener/resourcemanager/constants"
+	"github.com/gardener/gardener/pkg/component/extensions/operatingsystemconfig"
 	kubeapiserver "github.com/gardener/gardener/pkg/component/kubernetes/apiserver"
-	"github.com/gardener/gardener/pkg/controllerutils"
 	"github.com/gardener/gardener/pkg/gardenadm/staticpod"
+	"github.com/gardener/gardener/pkg/utils"
 	gardenerutils "github.com/gardener/gardener/pkg/utils/gardener"
 	kubernetesutils "github.com/gardener/gardener/pkg/utils/kubernetes"
 	secretsutils "github.com/gardener/gardener/pkg/utils/secrets"
@@ -41,11 +33,14 @@ import (
 // PathKubeconfig is the path to a file on the control plane node containing an admin kubeconfig.
 var PathKubeconfig = filepath.Join(string(filepath.Separator), "etc", "kubernetes", "admin.conf")
 
-// KubeconfigSecretName is the name of the secret in the shoot namespace of the bootstrap cluster containing the
-// kubeconfig of the self-hosted shoot.
-const KubeconfigSecretName = "kubeconfig"
+type staticControlPlaneComponent struct {
+	deploy       func(context.Context) error
+	name         string
+	targetObject client.Object
+	mutate       func(*corev1.Pod)
+}
 
-func (b *GardenadmBotanist) deployETCD(role string) func(context.Context) error {
+func (b *Botanist) deployETCD(role string) func(context.Context) error {
 	var portClient, portPeer, portMetrics int32 = 2379, 2380, 2381
 	if role == v1beta1constants.ETCDRoleEvents {
 		portClient, portPeer, portMetrics = etcdconstants.StaticPodPortEtcdEventsClient, 2383, 2384
@@ -67,45 +62,16 @@ func (b *GardenadmBotanist) deployETCD(role string) func(context.Context) error 
 	}
 }
 
-func (b *GardenadmBotanist) deployKubeAPIServer(ctx context.Context) error {
+func (b *Botanist) deployKubeAPIServer(ctx context.Context) error {
 	b.Shoot.Components.ControlPlane.KubeAPIServer.EnableStaticTokenKubeconfig()
 	b.Shoot.Components.ControlPlane.KubeAPIServer.SetAutoscalingReplicas(new(int32(0)))
-
-	// kube-apiserver must be able to resolve the gardener-resource-manager service IP to access the
-	// node-agent-authorizer webhook. Thus, we fetch the service IP. It is used to create a host alias in the
-	// kube-apiserver pod spec later. If the service does not exist yet, then kube-apiserver is bootstrapped for the
-	// first time - in this case, we don't activate the authorizer webhook.
-	service := &corev1.Service{ObjectMeta: metav1.ObjectMeta{Name: resourcemanagerconstants.ServiceName, Namespace: b.Shoot.ControlPlaneNamespace}}
-	if err := b.SeedClientSet.Client().Get(ctx, client.ObjectKeyFromObject(service), service); err != nil {
-		if !apierrors.IsNotFound(err) {
-			return fmt.Errorf("failed getting service %s: %w", client.ObjectKeyFromObject(service), err)
-		}
-	}
-
-	b.gardenerResourceManagerServiceIPs = service.Spec.ClusterIPs
-
 	return b.DeployKubeAPIServer(ctx)
 }
 
-type staticControlPlaneComponent struct {
-	deploy       func(context.Context) error
-	name         string
-	targetObject client.Object
-	mutate       func(*corev1.Pod)
-}
-
-func (b *GardenadmBotanist) staticControlPlaneComponents() []staticControlPlaneComponent {
+func (b *Botanist) staticControlPlaneComponents(useBootstrapEtcd bool) []staticControlPlaneComponent {
 	var (
 		components []staticControlPlaneComponent
 
-		mutateAPIServerPodFn = func(pod *corev1.Pod) {
-			for _, ip := range b.gardenerResourceManagerServiceIPs {
-				pod.Spec.HostAliases = append(pod.Spec.HostAliases, corev1.HostAlias{
-					IP:        ip,
-					Hostnames: []string{resourcemanagerconstants.ServiceName},
-				})
-			}
-		}
 		mutateETCDPodFn = func(pod *corev1.Pod) {
 			// The pod name of the etcd pod must match `etcd-<role>-0`. However, when running as static pod, its pod name is
 			// `etcd-<role>-<node-name>`. Hence, let's mutate it here for now (this might not be the final solution
@@ -117,7 +83,7 @@ func (b *GardenadmBotanist) staticControlPlaneComponents() []staticControlPlaneC
 		}
 	)
 
-	if !b.useEtcdManagedByDruid {
+	if useBootstrapEtcd {
 		components = append(components,
 			staticControlPlaneComponent{b.deployETCD(v1beta1constants.ETCDRoleMain), bootstrapetcd.Name(v1beta1constants.ETCDRoleMain), &appsv1.StatefulSet{}, nil},
 			staticControlPlaneComponent{b.deployETCD(v1beta1constants.ETCDRoleEvents), bootstrapetcd.Name(v1beta1constants.ETCDRoleEvents), &appsv1.StatefulSet{}, nil},
@@ -130,21 +96,21 @@ func (b *GardenadmBotanist) staticControlPlaneComponents() []staticControlPlaneC
 	}
 
 	return append(components,
-		staticControlPlaneComponent{b.deployKubeAPIServer, v1beta1constants.DeploymentNameKubeAPIServer, &appsv1.Deployment{}, mutateAPIServerPodFn},
+		staticControlPlaneComponent{b.deployKubeAPIServer, v1beta1constants.DeploymentNameKubeAPIServer, &appsv1.Deployment{}, nil},
 		staticControlPlaneComponent{b.DeployKubeControllerManager, v1beta1constants.DeploymentNameKubeControllerManager, &appsv1.Deployment{}, nil},
 		staticControlPlaneComponent{b.Shoot.Components.ControlPlane.KubeScheduler.Deploy, v1beta1constants.DeploymentNameKubeScheduler, &appsv1.Deployment{}, nil},
 	)
 }
 
-// DeployControlPlaneDeployments deploys the deployments for the static control plane components. It also updates the
-// OperatingSystemConfig, waits for it to be reconciled by the OS extension, and deploys the ManagedResource containing
-// the Secret with OperatingSystemConfig for gardener-node-agent.
-func (b *GardenadmBotanist) DeployControlPlaneDeployments(ctx context.Context) error {
-	if err := b.deployControlPlaneDeployments(ctx); err != nil {
+// DeployStaticControlPlaneDeployments deploys the deployments for the static control plane components. It also updates
+// the OperatingSystemConfig, waits for it to be reconciled by the OS extension, and deploys the ManagedResource
+// containing the Secret with OperatingSystemConfig for gardener-node-agent.
+func (b *Botanist) DeployStaticControlPlaneDeployments(ctx context.Context, useBootstrapEtcd bool) error {
+	if err := b.DeployControlPlaneDeployments(ctx, useBootstrapEtcd); err != nil {
 		return fmt.Errorf("failed deploying control plane deployments: %w", err)
 	}
 
-	if _, _, err := b.deployOperatingSystemConfig(ctx); err != nil {
+	if _, _, err := b.DeployOperatingSystemConfigWithStaticPods(ctx, useBootstrapEtcd); err != nil {
 		return fmt.Errorf("failed deploying OperatingSystemConfig: %w", err)
 	}
 
@@ -160,8 +126,9 @@ func (b *GardenadmBotanist) DeployControlPlaneDeployments(ctx context.Context) e
 	return nil
 }
 
-func (b *GardenadmBotanist) deployControlPlaneDeployments(ctx context.Context) error {
-	for _, component := range b.staticControlPlaneComponents() {
+// DeployControlPlaneDeployments runs the Deploy function of the control plane components.
+func (b *Botanist) DeployControlPlaneDeployments(ctx context.Context, useBootstrapEtcd bool) error {
+	for _, component := range b.staticControlPlaneComponents(useBootstrapEtcd) {
 		if err := b.deployControlPlaneComponent(ctx, component.deploy, component.targetObject, component.name); err != nil {
 			return fmt.Errorf("failed deploying %q: %w", component.name, err)
 		}
@@ -170,7 +137,7 @@ func (b *GardenadmBotanist) deployControlPlaneDeployments(ctx context.Context) e
 	return nil
 }
 
-func (b *GardenadmBotanist) deployControlPlaneComponent(ctx context.Context, deploy func(context.Context) error, targetObject client.Object, componentName string) error {
+func (b *Botanist) deployControlPlaneComponent(ctx context.Context, deploy func(context.Context) error, targetObject client.Object, componentName string) error {
 	if err := deploy(ctx); err != nil {
 		return fmt.Errorf("failed deploying component %q: %w", componentName, err)
 	}
@@ -184,7 +151,7 @@ func (b *GardenadmBotanist) deployControlPlaneComponent(ctx context.Context, dep
 	return b.SeedClientSet.Client().Get(ctx, client.ObjectKeyFromObject(targetObject), targetObject)
 }
 
-func (b *GardenadmBotanist) populateStaticAdminTokenToAccessTokenSecret(ctx context.Context, componentName string) error {
+func (b *Botanist) populateStaticAdminTokenToAccessTokenSecret(ctx context.Context, componentName string) error {
 	secret := &corev1.Secret{ObjectMeta: metav1.ObjectMeta{Name: gardenerutils.SecretNamePrefixShootAccess + componentName, Namespace: b.Shoot.ControlPlaneNamespace}}
 	if err := b.SeedClientSet.Client().Get(ctx, client.ObjectKeyFromObject(secret), secret); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -216,6 +183,56 @@ func (b *GardenadmBotanist) populateStaticAdminTokenToAccessTokenSecret(ctx cont
 	return b.SeedClientSet.Client().Update(ctx, secret)
 }
 
+// DeployOperatingSystemConfigWithStaticPods deploys the OperatingSystemConfig containing the files for the control
+// plane components running as static pods.
+func (b *Botanist) DeployOperatingSystemConfigWithStaticPods(ctx context.Context, useBootstrapEtcd bool) (*operatingsystemconfig.Data, string, error) {
+	pods, err := b.staticControlPlanePods(ctx, useBootstrapEtcd)
+	if err != nil {
+		return nil, "", fmt.Errorf("failed computing files for static control plane pods: %w", err)
+	}
+
+	files, err := b.appendAdminKubeconfigToFiles(pods.allFiles())
+	if err != nil {
+		return nil, "", fmt.Errorf("failed appending admin kubeconfig to list of files: %w", err)
+	}
+
+	if err := b.DeployOperatingSystemConfig(ctx); err != nil {
+		return nil, "", fmt.Errorf("failed deploying OperatingSystemConfig resource: %w", err)
+	}
+
+	controlPlaneWorkerPool := v1beta1helper.ControlPlaneWorkerPoolForShoot(b.Shoot.GetInfo().Spec.Provider.Workers)
+	if controlPlaneWorkerPool == nil {
+		return nil, "", fmt.Errorf("failed fetching the control plane worker pool for the shoot")
+	}
+
+	oscData, ok := b.Shoot.Components.Extensions.OperatingSystemConfig.WorkerPoolNameToOperatingSystemConfigsMap()[controlPlaneWorkerPool.Name]
+	if !ok {
+		return nil, "", fmt.Errorf("failed fetching the generated OperatingSystemConfig data for the control plane worker pool %q", controlPlaneWorkerPool.Name)
+	}
+	osc := oscData.Original.Object
+
+	patch := client.MergeFrom(osc.DeepCopy())
+	osc.Spec.Files = append(osc.Spec.Files, files...)
+	if err := b.SeedClientSet.Client().Patch(ctx, osc, patch); err != nil {
+		return nil, "", fmt.Errorf("failed patching OperatingSystemConfig with additional files for static control plane pods: %w", err)
+	}
+
+	return &oscData.Original, controlPlaneWorkerPool.Name, nil
+}
+
+func (b *Botanist) appendAdminKubeconfigToFiles(files []extensionsv1alpha1.File) ([]extensionsv1alpha1.File, error) {
+	userKubeconfigSecret, ok := b.SecretsManager.Get(kubeapiserver.SecretNameUserKubeconfig)
+	if !ok {
+		return nil, fmt.Errorf("failed fetching secret %q", kubeapiserver.SecretNameUserKubeconfig)
+	}
+
+	return append(files, extensionsv1alpha1.File{
+		Path:        PathKubeconfig,
+		Permissions: new(uint32(0600)),
+		Content:     extensionsv1alpha1.FileContent{Inline: &extensionsv1alpha1.FileContentInline{Encoding: "b64", Data: utils.EncodeBase64(userKubeconfigSecret.Data[secretsutils.DataKeyKubeconfig])}},
+	}), nil
+}
+
 type staticPod struct {
 	name  string
 	files []extensionsv1alpha1.File
@@ -231,10 +248,10 @@ func (s staticPods) allFiles() []extensionsv1alpha1.File {
 	return files
 }
 
-func (b *GardenadmBotanist) staticControlPlanePods(ctx context.Context) (staticPods, error) {
+func (b *Botanist) staticControlPlanePods(ctx context.Context, useBootstrapEtcd bool) (staticPods, error) {
 	var pods staticPods
 
-	for _, component := range b.staticControlPlaneComponents() {
+	for _, component := range b.staticControlPlaneComponents(useBootstrapEtcd) {
 		if err := b.SeedClientSet.Client().Get(ctx, client.ObjectKey{Name: component.name, Namespace: b.Shoot.ControlPlaneNamespace}, component.targetObject); err != nil {
 			return nil, fmt.Errorf("failed reading object for %q: %w", component.name, err)
 		}
@@ -251,93 +268,4 @@ func (b *GardenadmBotanist) staticControlPlanePods(ctx context.Context) (staticP
 	}
 
 	return pods, nil
-}
-
-// NewClientSetFromFile creates a client set from the specified kubeconfig file.
-func NewClientSetFromFile(kubeconfigPath string, scheme *runtime.Scheme) (kubernetes.Interface, error) {
-	return kubernetes.NewClientFromFile("", kubeconfigPath,
-		kubernetes.WithClientOptions(client.Options{Scheme: scheme}),
-		kubernetes.WithClientConnectionOptions(componentbaseconfigv1alpha1.ClientConnectionConfiguration{QPS: 100, Burst: 130}),
-		kubernetes.WithDisabledCachedClient(),
-	)
-}
-
-// CreateClientSet creates a client set for the control plane.
-func (b *GardenadmBotanist) CreateClientSet(ctx context.Context) (kubernetes.Interface, error) {
-	pathKubeconfig := PathKubeconfig
-	if path := os.Getenv("KUBECONFIG"); path != "" {
-		pathKubeconfig = path
-	}
-
-	clientSet, err := NewClientSetFromFile(pathKubeconfig, kubernetes.SeedScheme)
-	if err != nil {
-		b.Logger.Info("Waiting for kube-apiserver to start", "error", err.Error())
-		return nil, fmt.Errorf("failed creating client set: %w", err)
-	}
-
-	clientSet.Start(ctx)
-
-	waitContext, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-	if !clientSet.WaitForCacheSync(waitContext) {
-		return nil, fmt.Errorf("timed out waiting for caches")
-	}
-
-	result := clientSet.RESTClient().Get().AbsPath("/readyz").Do(ctx)
-	if result.Error() != nil {
-		return nil, fmt.Errorf("failed to GET /readyz endpoint of kube-apiserver: %w", result.Error())
-	}
-
-	var statusCode int
-	result.StatusCode(&statusCode)
-	if statusCode != http.StatusOK {
-		b.Logger.Info("The kube-apiserver does not report readiness yet", "statusCode", statusCode)
-		return nil, fmt.Errorf("kube-apiserver does not report readiness yet")
-	}
-
-	return clientSet, nil
-}
-
-// DiscoverKubernetesVersion discovers the Kubernetes version of the control plane.
-func (b *GardenadmBotanist) DiscoverKubernetesVersion(clientSet kubernetes.Interface) (*semver.Version, error) {
-	version, err := semver.NewVersion(clientSet.Version())
-	if err != nil {
-		return nil, fmt.Errorf("failed parsing semver version %q: %w", clientSet.Version(), err)
-	}
-
-	return version, nil
-}
-
-// FetchKubeconfig fetches the kubeconfig of the self-hosted shoot from the control plane machine via SSH and stores it
-// in a secret in the shoot namespace of the bootstrap cluster for later retrieval. It also writes the kubeconfig to the
-// given output writer, if any.
-func (b *GardenadmBotanist) FetchKubeconfig(ctx context.Context, output io.Writer) error {
-	kubeconfigBuffer := &bytes.Buffer{}
-	if err := b.sshConnection.SCP.CopyFromRemotePassThru(ctx, kubeconfigBuffer, PathKubeconfig, nil); err != nil {
-		return fmt.Errorf("error fetching kubeconfig: %w", err)
-	}
-	kubeconfigBytes := kubeconfigBuffer.Bytes()
-
-	kubeconfigSecret := &corev1.Secret{ObjectMeta: metav1.ObjectMeta{Name: KubeconfigSecretName, Namespace: b.Shoot.ControlPlaneNamespace}}
-	_, err := controllerutils.CreateOrGetAndMergePatch(ctx, b.SeedClientSet.Client(), kubeconfigSecret, func() error {
-		kubeconfigSecret.Data = map[string][]byte{
-			secretsutils.DataKeyKubeconfig: kubeconfigBytes,
-		}
-		return nil
-	})
-	if err != nil {
-		return fmt.Errorf("error writing kubeconfig secret: %w", err)
-	}
-	b.Logger.Info("Stored kubeconfig of the self-hosted shoot in secret", "namespace", kubeconfigSecret.Namespace, "name", kubeconfigSecret.Name)
-
-	if output != nil {
-		if outputFile, ok := output.(*os.File); ok {
-			b.Logger.Info("Writing kubeconfig of the self-hosted shoot to file", "path", outputFile.Name())
-		}
-		if _, err := output.Write(kubeconfigBytes); err != nil {
-			return fmt.Errorf("error writing kubeconfig to output: %w", err)
-		}
-	}
-
-	return nil
 }
